% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioticExplorerServer-package.R
\docType{package}
\name{BioticExplorerServer-package}
\alias{BioticExplorerServer}
\alias{BioticExplorerServer-package}
\title{R Package for Downloading Server-Side Data for BioticExplorer}
\description{
This package can be used to download IMR Biotic database and to place it into
 a \link[duckdb]{duckdb} database, providing access to all of the institute’s
 Biotic data from R.
}
\details{
The database can be downloaded and compiled using the \code{\link{compileDatabase}} function.
}
\author{
\strong{Maintainer}: Mikko Vihtakari \email{mikko.vihtakari@hi.no}

Authors:
\itemize{
  \item Ibrahim Umar
  \item The Stox Project team
}

Other contributors:
\itemize{
  \item Institute of Marine Research [copyright holder]
}

}
\keyword{internal}
