% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbPath.R
\name{dbPath}
\alias{dbPath}
\title{The file file path to the MonetDB folder}
\usage{
dbPath()
}
\description{
This function exports the file path to the MonetDB database folder. The path has to be changed to fit the server.
}
