% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{ICESareas}
\alias{ICESareas}
\title{ICES fishing areas}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} in decimal degrees (+init=epsg:4326).
}
\source{
\href{https://gis.ices.dk/sf/index.html}{International Council for the Exploration of the Sea}
}
\usage{
ICESareas
}
\description{
Food and Agriculture Organization Major Fishing Area 27 (i.e. ICES region) fishing area polygons. The polygons are not cut with land as opposed to the areas distributed on the ICES website (see \code{BioticEXplorerServer:::prepareICESareas}). This makes the polygons smaller and consequent calculations quicker.
}
\concept{shapefiles}
\keyword{datasets}
\keyword{shapefiles}
