% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexDatabase.R
\name{indexDatabase}
\alias{indexDatabase}
\title{Index BioticExplorer database}
\usage{
indexDatabase(
  connection,
  dbIndexFile = file.path(defaultDbPath(), "dbIndex.rda"),
  fileOnly = TRUE
)
}
\arguments{
\item{connection}{Object defining the \link[duckdb]{duckdb} connection. Typically made within \code{\link{compileDatabase}}.}

\item{dbIndexFile}{Character string specifying the file path where the index of the database should be saved. Must include \code{.rda} at the end. Defaults to \code{dbIndex.rda} inside \code{dbPath}. The index is used by \href{https://github.com/DeepWaterIMR/BioticExplorer}{BioticExplorer}.}

\item{fileOnly}{Logical indicating whether the result should only be saved to a file and not returned. If FALSE, no file is made and the result is returned instead.}
}
\value{
If \code{fileOnly = TRUE} (the default), saves the index to \code{dbIndexFile} and returns \code{NULL} invisibly. Otherwise returns a named list of pre-computed index values used by the BioticExplorer Shiny app.
}
\description{
Loads BioticExplorer database and creates an index used by BioticExplorer to save processing time.
}
\author{
Mikko Vihtakari, Ibrahim Umar (Institute of Marine Research)
}
