% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareICESareas.R
\name{prepareICESareas}
\alias{prepareICESareas}
\title{Download and format ICES area polygons}
\usage{
prepareICESareas()
}
\description{
Download and format ICES area polygons
}
\details{
This function is an Ibrahim hack which does not work optimally because the area polygons supplied on the ICES site are very complex and partly corrupted. Use the premade polygons (\code{data(icesAreas)}) instead. These polygons have been obtained from the ICES GIS team upon request and much simpler.
}
\author{
Ibrahim Umar
}
\keyword{internal}
