% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioticToDatabase.R
\name{bioticToDatabase}
\alias{bioticToDatabase}
\title{Read and process a NMD Biotic xml file for further use in the BioticExplorer database}
\usage{
bioticToDatabase(
  file,
  removeEmpty = FALSE,
  convertColumns = TRUE,
  missionidPrefix = NULL,
  icesAreaShape = NULL,
  cruiseSeries
)
}
\arguments{
\item{file}{character string specifying the file path to the xml file. Accepts only one file at the time.}

\item{removeEmpty}{logical indicating whether empty columns should be removed from the output. This option also influences "coreData" columns.}

\item{convertColumns}{logical indicating whether the column types should be converted. See \code{link{convertColumnTypes}}. Setting this to \code{FALSE} considerably speeds up the function, but leads to problems with non-unicode characters.}

\item{missionidPrefix}{A prefix for the \code{missionid} identifier, which separates cruises. \code{NULL} (default) omits the prefix. Use year when writing to the database.}

\item{icesAreaShape}{ICES area shape in SpatialPolygonsDataFrame object. Used for calculating the ICES area for a specific fishstation.}

\item{cruiseSeries}{a data.table object of NMD cruise series list. Used to identify cruise series of a specific mission.}
}
\value{
Returns a list of Biotic data with the \code{$mission} data table from the original NMD data. The \code{$stnall} and \code{$indall} data frames are merged from \code{$fishstation} and \code{$catchsample} (former) and  \code{$fishstation}, \code{$catchsample}, \code{$individual} and \code{$agedetermination} (latter).
}
\description{
A wrapper for \code{\link[RstoxData]{readXmlFile}} to enable further use in the BioticExplorer database
}
\author{
Mikko Vihtakari, Ibrahim Umar (Institute of Marine Research)
}
