% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileDatabase.R
\name{compileDatabase}
\alias{compileDatabase}
\title{Compile BioticExplorer database}
\usage{
compileDatabase(
  years = 1900:data.table::year(Sys.time()),
  dbPath = "~/Documents/IMR_biotic_BES_database",
  dbIndexFile = file.path(dbPath, "dbIndex.rda"),
  dbName = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{years}{vector of integer specifying the years to be downloaded. The database reaches 1914:2020}

\item{dbName}{Character string or \code{NULL}. If \code{NULL} uses the default names and overwrites the existing database.}

\item{overwrite}{Logical indicating whether existing information in the \code{dbName} should be downloaded again and overwritten.}

\item{dbIndexPath}{Character string specifying the file path where the database should be saved. Must include \code{.rda} at the end.}
}
\description{
Downloads, formulates and indexes BioticExplorer database.
}
\details{
Runs the \code{\link{downloadDatabase}} and \code{\link{indexDatabase}} functions. Be aware that running these functions requires access to the IMR intranet, reasonably fast internet and loads of memory. If the function crashes after the \code{\link{downloadDatabase}}, you can still run the \code{\link{indexDatabase}} to save the progress. If it crashes during \code{\link{downloadDatabase}}, you may have to start from scratch.
}
\author{
Mikko Vihtakari, Ibrahim Umar (Institute of Marine Research)
}
