% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileDatabase.R
\name{compileDatabase}
\alias{compileDatabase}
\title{Compile BioticExplorer database}
\usage{
compileDatabase(
  years = 1900:data.table::year(Sys.time()),
  dbIndexPath = "~/Desktop/dbIndex.rda",
  source = NULL,
  dbName = NULL,
  prepareCruiseSeries = TRUE
)
}
\arguments{
\item{years}{vector of integer specifying the years to be downloaded. The database reaches 1914:2020}

\item{dbIndexPath}{Character string specifying the file path where the database should be saved. Must include \code{.rda} at the end.}

\item{source}{Character string indicating from where the database should be compiled. If \code{NULL}, the data will be downloaded from NMD. Requires access to the IMR intranet. Otherwise, specify a file path where the rds files are located from the \code{\link{downloadDatabaseToFiles}} function. NOT IMPLEMENTED YET.}

\item{dbName}{Character string or \code{NULL}. If \code{NULL} uses the default names and overwrites the existing database.}

\item{prepareCruiseSeries}{Logical indicating whether \link[=prepareCruiseSeriesList]{the cruise series list} should be prepared. Only set this to \code{FALSE} while debugging as preparing the cruise series list takes time.}
}
\description{
Downloads, formulates and indexes BioticExplorer database.
}
\details{
Runs the \code{\link{downloadDatabase}} and \code{\link{indexDatabase}} functions. Be aware that running these functions requires access to the IMR intranet, reasonably fast internet and loads of memory. If the function crashes after the \code{\link{downloadDatabase}}, you can still run the \code{\link{indexDatabase}} to save the progress. If it crashes during \code{\link{downloadDatabase}}, you may have to start from scratch.
}
\author{
Mikko Vihtakari, Ibrahim Umar (Institute of Marine Research)
}
