% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDatabase.R
\name{downloadDatabase}
\alias{downloadDatabase}
\title{Download and parse NMD data for the BioticExplorer database}
\usage{
downloadDatabase(
  years,
  connection,
  icesAreas = icesAreas,
  cruiseSeries = cruiseSeries,
  gearCodes = gearCodes,
  overwrite = FALSE
)
}
\arguments{
\item{years}{Vector of integers specifying the years to be downloaded. The database reaches 1914:year(Sys.Date())}

\item{connection}{Object defining the \link[duckdb]{duckdb} connection. Typically made within \code{\link{compileDatabase}}.}

\item{icesAreas}{ICES area shape \code{\link[sf]{st_polygon}} abject. Used for calculating the ICES area for a specific fishstation.}

\item{cruiseSeries}{a data.table object of NMD cruise series list. Used to identify cruise series of a specific mission. See \code{\link{prepareCruiseSeriesList}}.}

\item{gearCodes}{a data.table object of NMD gear code list. Used to make gearname and gearcategory columns. See \code{\link{prepareGearList}}.}

\item{overwrite}{Logical indicating whether existing information in the \link[duckdb]{duckdb} database (\code{dbPath}) should be downloaded again and overwritten.}
}
\description{
Downloads annual NMD data from the API and writes them as a DuckDB database
}
\details{
The function downloads NMD data from the API per year, saves these in temp files, reformats them for the DuckDB and writes them into the database.
}
\author{
Ibrahim Umar, Mikko Vihtakari (Institute of Marine Research)
}
