% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDatabase.R
\name{downloadDatabase}
\alias{downloadDatabase}
\title{Download and parse NMD data for the BioticExplorer database}
\usage{
downloadDatabase(years, dbPath = dbPath())
}
\arguments{
\item{years}{vector of integer specifying the years to be downloaded. The database reaches 1900:2020}

\item{dbPath}{Character string specifying the file path where the database should be located. Must include \code{.monetdb} at the end.}
}
\description{
Downloads annual NMD data from the API and writes them as MonetDB database
}
\details{
This function is scarily powerful. Do not run a large number of years unless you think you know what you are doing
}
\author{
Ibrahim Umar, Mikko Vihtakari (Institute of Marine Research)
}
