% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexDatabase.R
\name{indexDatabase}
\alias{indexDatabase}
\title{Index BioticExplorer database}
\usage{
indexDatabase(
  dbIndexPath = "~/Desktop/dbIndex.rda",
  fileOnly = TRUE,
  dbName = NULL
)
}
\arguments{
\item{dbIndexPath}{Character string specifying the file path where the database should be saved. Must include \code{.rda} at the end.}

\item{fileOnly}{Logical indicating whether the result should only be saved to a file and not returned. If FALSE, no file is made and the result is returned instead.}

\item{dbName}{Character string or \code{NULL}. If \code{NULL} uses the default names and overwrites the existing database.}
}
\description{
Loads BioticExplorer database and creates an index used by BioticExplorer to save processing time
}
\author{
Mikko Vihtakari, Ibrahim Umar (Institute of Marine Research)
}
