% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTaxaList.R
\name{prepareTaxaList}
\alias{prepareTaxaList}
\title{Prepare taxa list}
\usage{
prepareTaxaList(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical indicating whether the function should return status information as messages.}
}
\description{
Prepares taxa list using the IMR database API. Stored as data object in the package. Refreshes any new taxa codes.
}
\author{
Mikko Vihtakari, Ibrahim Umar
}
